restructure and adding mantine
