+++ src/App.jsx
@@ -1,34 +1,34 @@
-import { Routes, Route, } from 'react-router-dom';
+import {Route, Routes,} from 'react-router-dom';
 
-import RequireAuth from './components/RequireAuth';
+import RequireAuth from './old/components/RequireAuth';
 
-import TopMenu from './components/TopMenu';
-import Login from './components/Login';
-import Home from './pages/home';
-import PG from './pages/PG';
+import TopMenu from './old/components/TopMenu';
+import Login from './pages/Login';
+import Home from './old/pages/home';
+import PG from './old/pages/PG';
 import Test from './Test';
 
 function App() {
-  return (
-    <Routes>
-      <Route path="/">
-        {/*public routes*/}
-        <Route path="login" element={<Login />} />
+    return (
+        <Routes>
+            <Route path="/">
+                {/*public routes*/}
+                <Route path="login" element={<Login/>}/>
 
-        {/*privates routes*/}
-        <Route element={<RequireAuth />}>
-          <Route path="/" element={<TopMenu />}>
-            <Route path="/" element={<Home />} />
-            <Route path="pg/:pgId" element={<PG />} />
-            <Route path="test" element={<Test />} />
-          </Route>
-        </Route>
+                {/*privates routes*/}
+                <Route element={<RequireAuth/>}>
+                    <Route path="/" element={<TopMenu/>}>
+                        <Route path="/" element={<Home/>}/>
+                        <Route path="pg/:pgId" element={<PG/>}/>
+                        <Route path="test" element={<Test/>}/>
+                    </Route>
+                </Route>
 
-        {/* 404*/}
-        {/* TODO:*/}
-      </Route>
-    </Routes >
-  );
+                {/* 404*/}
+                {/* TODO:*/}
+            </Route>
+        </Routes>
+    );
 }
 
 export default App;
