+++ src/index.js
@@ -1,16 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import 'bootstrap/dist/css/bootstrap.min.css';
+import {MantineProvider} from '@mantine/core';
 import App from './App';
 
-import { BrowserRouter, Routes, Route, } from 'react-router-dom';
+import {BrowserRouter, Route, Routes,} from 'react-router-dom';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <BrowserRouter>
-    <Routes>
-      <Route path="/*" element={<App />} />
-    </Routes>
-  </BrowserRouter>
+    <MantineProvider withGlobalStyles withNormalizeCSS>
+        <BrowserRouter>
+            <Routes>
+                <Route path="/*" element={<App/>}/>s
+            </Routes>
+        </BrowserRouter>
+    </MantineProvider>
 );
